getFirstParam, nullFill
Added getFirstParam procedure. Need to adjust it more for whitespace, but I'll do that later. It takes the next part of the string (passed to it in edi) and copies the following string after a space. Also added nullFill which fills an array with 0s. I used it to wipe the firstParam variable used to hold the first parameter.