data structure
Added the skeleton for the data structure. Fixed some of the errors with data entry but not all. Started working on the cmdShow procedure. Have the name part done. Kind of need to get the cmdLoad procedure up before I can test cmdShow first though. There's probably errors in it already.