String update and misc
The bug in the cmpString procedure was not a bug, I had to check if the null character was being encourntered on the string literal before the comparison on the two characters was occuring, this allowed for the program to check for the part of a string and let us continue to pass parameters to the string. Also, I made the print and println macros transient. And played around with the load and show procedures. Still really no success though.