change working
Fixed the change function. The problem was that the nullFill macro was overwriting the data. So i reworked it since I have a better understanding of pointers now. Made it into a procedure. It's all documented. Now all that's left is step. And to work out some kinks in the chain.